This could be called a final draft even though it's pretty bad.
This could REALLY use improvement, mostly on the last part of the code. I need to make it so that after the use puts in the value to checkyoself, the code doesn't finish, and you can keep on checking your self without having to re-input all of the code all over again.