Update Flashcardcreation
Created a function to create a new set of flashcards.  It takes a list of topics as input, iterates through each topic in the list, and adds that topic to the dictionary.  The user can then enter what they want for the back of the card (what they want to say about the topic).