Create JoshJongwonLee
I fixed it! Now it iterates over and over until the user types quit. 
The program terminates when the user types quit. Plus, I noticed that when the user 
initially types in the front parts and back parts of the card, there should be no space 
in between at all. So I added a notification saying that there should be no space. 
Plus, when every the program iterates, it shows the list of the front part again. 
It is much more convenient this way so that the user can look at every time he/she wants 
to type another front part - just like a word bank on a vocabulary test. 
I also added some lines just so that it looks neater.